server/requirements.txt
flask
flask-bcrypt
python-dotenv
flask-session
flask-cors
redis